Tune up webpacker
Introduce new `lcms_engine_javascript_pack_tag` method to be used to include(and compile) lcms-engine based packs.

This is a solution to be able to isolate project specific packs and packs created inside the gem.

As an alternative we could use the solution provided here - https://github.com/rails/webpacker/issues/348#issuecomment-635480949

This looks more reliable to me.

- Update Dockerfile to set explicitly locale.