Allow `ImportStatus` page customization
Mostly the whole `ImportStatus` component came from a legacy code base. Now, when we widely use it across different projects it might me useful to have a way to customize it a bit.

Now, when preparing the `@props` in the host app, we can use additional parameter `text` which will be shown during the job. If it's omitted - the value of the `link` will be used.

```ruby
jobs: { job_id => { link: folder_url, status: 'waiting', text: 'Copying...' } }
```