Enhance error handling for job result fetching
Improved error reporting in the 'reimportable' concern by extending the condition to handle the case when 'jid_res' is not present. This change ensures that informative error messages are returned for all possible error cases, thus improving debugging efficiency.

Prevents an exception when UI is trying to fetch the job information but job has already been deleted.