Remove unnecessary priority icon and metadata injection
This commit includes deletion of priority icons and forced injection of metadata in the code. The priority_icon method was used for displaying priority icons in the document which are no longer required, making this code redundant. Similarly, force_inject_section method was used for injecting the Section metadata only if there are no Agenda tables which is not needed anymore. The changes were also reflected in the html templates where priority_icon was used. So their references were also removed. These changes will help clean unnecessary code and improve readability.