Fix typo in Resource and refactor code to improve readability and performance
The base_tag.rb and resource.rb files were refactored to
improve both the readability of the code and the performance of the functions. In base_tag.rb, redundant 'self' keywords were removed from the 'const_defined?' method calls in 'tag_with_html_regexp' and 'tag_with_html_regexp_array' functions, making the code cleaner. In resource.rb, the method call inside 'where' method in 'with_curriculum' function was modified to directly access the id of the default Curriculum value, improving the complexity and efficiency of this query.