Remove unused file, update certain methods and create new RBS signatures
This commit does several things:

- Deletes the unused `rubyzip.rbs` file. Its signature are now taken from rbs_collection gem
- Adds the `Scopes` module to `standard.rbs` which includes two new methods, `by_grade` and `by_grades`.
- Creates new files, `base_query.rbs`, `admin_materials_query.rbs`, and `admin_documents_query.rbs` all of which define new classes for querying specific entities.
- Modifies some method definitions in `standard.rb`, `base_tag.rb`, and `document_bundle.rb`.
- Updates the rbs_collection
- Creates new files, `csv.rbs` and `bundler.rbs` for handling certain operations.

These changes are made for better code organization and improved query handling. The updates to the various method definitions are necessary to ensure that they operate correctly under different conditions. Moreover, the `rbs_collection.lock.yaml` file is updated to reflect the current state of the system's dependencies.