Correctly get the constant values from within child classes
The TAG_NAME constant is now fetched using const_get method to accommodate for derived classes where the constant might have a different value than the base class. It makes the code more extendable for future modifications. The redundant TAG_NAME in file base_tag.rbs was also removed, as it was already present in ChildrenConstants.