Add simple bold, italic, line break tags support
This update adds a new function, `substitute_tags_in`, to `DocTemplate::Tags`. This function takes in a string and returns the string with custom tags substituted with their corresponding HTML markup such as [b] for bold, [i] for italic, and [line-break] for line break. Corresponding unit tests are also included.